+++ CHANGELOG.md
@@ -10,6 +10,7 @@ next [????.??.??]
 * Add support for `kissat` and the `lingeling` trio (`lingeling`, `plingeling`,
   `treengeling`) of SAT solvers.
 * Add QBF examples (requires DepQBF solver)
+* Replace `test-framework` with `tasty` in the test suite.
 
 0.5 [2023.09.08]
 ----------------